tools/public-release-post-draft/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
openai>=1.51.0
pyyaml>=6.0